test: Rearranging the derivation engine tests
